export logo props
