+++ packages/logo/src/Logo.types.ts
@@ -1,4 +1,3 @@
-import { css } from '@leafygreen-ui/emotion';
 import { HTMLElementProps } from '@leafygreen-ui/lib';
 import { palette } from '@leafygreen-ui/palette';
 
@@ -23,7 +22,7 @@ type SupportedColorsMap =
 
 export { SupportedColors, SupportedColorsMap };
 
-export interface LogoProps extends HTMLElementProps<'svg'> {
+export interface ILogoProps extends HTMLElementProps<'svg'> {
   /**
    * Determines Color of the Logo or LogoMark component.
    *
@@ -46,57 +45,23 @@ export type ProductLogoProps = HTMLElementProps<'svg', never> & {
   height?: number;
 };
 
-export const getAccessibleProps = ({
-  role,
-  'aria-label': ariaLabel,
-}: {
-  role: string;
-  'aria-label': string;
-}) => {
-  if (role === 'img') {
-    return {
-      role: 'img',
-      'aria-label': ariaLabel,
-    };
-  }
+export const LogoNames = [
+  'MongoDBLogoMark',
+  'MongoDBLogo',
+  'AtlasNavGraphic',
+  'AtlasLogoLockup',
+  'AtlasForGovernmentLogoLockup',
+  'EnterpriseAdvancedLogoLockup',
+  'CommunityEditionLogoLockup',
+  'UniversityLogoLockup',
+] as const;
 
-  return {
-    role: 'presentation',
-    alt: '',
-    'aria-hidden': true,
-  };
-};
+export type LogoName = (typeof LogoNames)[number];
 
-export const getColor = ({
-  knockout,
-  size,
-  flat,
-  gradient,
-  darkMode,
-}: {
-  knockout: boolean;
-  darkMode: boolean;
-  size: number;
-  flat: string;
-  gradient: string;
-}) => {
-  let fill;
-
-  if (!knockout) {
-    if (size <= 10) {
-      fill = flat;
-    } else {
-      fill = gradient;
-    }
-  } else {
-    if (darkMode) {
-      fill = palette.white;
-    } else {
-      fill = palette.black;
-    }
-  }
-
-  return css`
-    fill: ${fill};
-  `;
-};
+export interface LogoProps extends ILogoProps {
+  /**
+   * The name of the logo to render
+   * @required
+   */
+  name: LogoName;
+}
