+++ packages/logo/src/Logos/AtlasLogoLockup.tsx
@@ -3,12 +3,8 @@ import PropTypes from 'prop-types';
 
 import { css, cx } from '@leafygreen-ui/emotion';
 
-import {
-  getAccessibleProps,
-  LogoProps,
-  SupportedColors,
-  SupportedColorsMap,
-} from './utils';
+import { ILogoProps, SupportedColors, SupportedColorsMap } from '../Logo.types';
+import { getAccessibleProps } from '../utils';
 
 /**
  * # AtlasLogoLockup
@@ -32,7 +28,7 @@ export const AtlasLogoLockup = React.forwardRef(
       role = 'img',
       'aria-label': ariaLabel = 'MongoDB Logo',
       ...rest
-    }: LogoProps,
+    }: ILogoProps,
     ref: React.LegacyRef<SVGSVGElement> | undefined,
   ): ReactElement => {
     const fill = SupportedColorsMap[color];
